smart-clock/requirements.txt
pygame
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
